requirements.txt
flask
flask-cors
numpy
pandas
astropy
torch
torchvision
matplotlib
pillow
pyyaml
tqdm